requirements.txt
numpy
torch
torchvision
torchaudio
matplotlib
tqdm
pandas
nibabel
pyYAML
black
